LLM_service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
aiohttp==3.9.3
sentence-transformers==2.2.2
psycopg[binary]==3.1.18
sqlalchemy[asyncio]==2.0.29
pgvector==0.2.4